Refactor icon drawing

+++ include/parts/icons.hpp
@@ -26,5 +26,5 @@
 
 namespace katzen {
 void drawIcon(
-    uint8_t iconId, float posX, float posY, uint8_t pixelSize, Color color);
+    uint8_t iconId, float posX, float posY, float pixelSize, Color color);
 } // namespace katzen
